% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{theme_ak}
\alias{theme_ak}
\title{Koh's default theme specs for {ggplot2}, inspired by functions in {silgelib}}
\usage{
theme_ak(
  base_size = 11,
  strip_text_size = 12,
  strip_text_margin = 5,
  subtitle_size = 13,
  subtitle_margin = 10,
  plot_title_size = 16,
  plot_title_margin = 10,
  legend_position = "bottom",
  plot_caption_color = "gray40",
  ...
)
}
\arguments{
\item{base_size}{base font size}

\item{strip_text_size, strip_text_margin}{plot strip text size and margin}

\item{subtitle_size, subtitle_margin}{plot subtitle size and margin}

\item{plot_title_size, plot_title_margin}{plot title size and margin}

\item{legend_position}{legend position}

\item{plot_caption_color}{color of caption}

\item{...}{Other arguments passed to \code{theme_minimal}}
}
\description{
Koh's default theme specs for {ggplot2}, inspired by functions in {silgelib}
}
\details{
The Roboto Condensed and Roboto Bold fonts are both Google fonts;
they can be found at \url{https://fonts.google.com/specimen/Roboto+Condensed}
and \url{https://fonts.google.com/specimen/Roboto}. These fonts must be
installed locally on your computer for this theme to work.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    labs(title = "A Lovely Plot",
         subtitle = "What can the subtitle tell us?") +
    theme_oswald()
}

}
